requirements.txt
# CLI Framework
typer[all]>=0.9.0

# Rich terminal output
rich>=13.0.0

# HTTP Requests
requests>=2.31.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Date handling
python-dateutil>=2.8.2

# Data validation
pydantic>=2.0.0

# Data processing (required by mite_client)
pandas>=2.0.0

# PDF generation
reportlab>=4.0.0
